backend/requirements.txt
flask
flask-cors
Pillow
torch
transformers
googletrans==4.0.0-rc1
